MB-31142 Fix for goroutines leak after undeploy
* Using single channel to control the  graceful exit of all
the goroutines in eventing consumer.

*Replaced channel with circular buffer in bounded queue.

Reviewed-on: http://review.couchbase.org/99395
Tested-by: Abhishek Singh <abhishek@couchbase.com>

+++ consumer/exported_functions.go
@@ -740,7 +740,7 @@ func (c *Consumer) UpdateWorkerQueueMemCap(quota int64) {
 
 	if c.app.UsingTimer {
 		divisor = 5
-		c.timerQueueMemCap = (quota / divisor) * 1024 * 1024
+		c.timerQueueMemCap = uint64((quota / divisor) * 1024 * 1024)
 	} else {
 		divisor = 2
 	}
